Fix pep8

+++ 0x0E-python-object_relational_mapping/13-model_state_delete_a.py
@@ -12,6 +12,7 @@ if __name__ == '__main__':
     Base.metadata.create_all(database)
     session_fake = sqlalchemy.orm.sessionmaker(bind=database)
     session = session_fake()
-    for element in session.query(State).filter(State.name.like('%a%')).order_by(State.id):
-        session.delete(element)
+    for element in session.query(State).order_by(State.id):
+        if 'a' in element.name:
+            session.delete(element)
     session.commit()
